Move complete Hook classes to their own modules
