+++ cloudbot/hooks/hook.py
@@ -1,0 +1,68 @@
+import asyncio
+import inspect
+import logging
+
+from cloudbot.hook import Action, Priority
+
+logger = logging.getLogger("cloudbot")
+
+
+class Hook:
+    """
+    Each hook is specific to one function. This class is never used by itself, rather extended.
+
+    :type type; str
+    :type plugin: Plugin
+    :type function: callable
+    :type function_name: str
+    :type required_args: list[str]
+    :type threaded: bool
+    :type permissions: list[str]
+    :type single_thread: bool
+    """
+
+    def __init__(self, _type, plugin, func_hook):
+        """
+        :type _type: str
+        :type plugin: Plugin
+        :type func_hook: hook._Hook
+        """
+        self.type = _type
+        self.plugin = plugin
+        self.function = func_hook.function
+        self.function_name = self.function.__name__
+
+        sig = inspect.signature(self.function)
+
+        # don't process args starting with "_"
+        self.required_args = [arg for arg in sig.parameters.keys() if not arg.startswith('_')]
+
+        if asyncio.iscoroutine(self.function) or asyncio.iscoroutinefunction(self.function):
+            self.threaded = False
+        else:
+            self.threaded = True
+
+        self.permissions = func_hook.kwargs.pop("permissions", [])
+        self.single_thread = func_hook.kwargs.pop("singlethread", False)
+        self.action = func_hook.kwargs.pop("action", Action.CONTINUE)
+        self.priority = func_hook.kwargs.pop("priority", Priority.NORMAL)
+
+        clients = func_hook.kwargs.pop("clients", [])
+
+        if isinstance(clients, str):
+            clients = [clients]
+
+        self.clients = clients
+
+        if func_hook.kwargs:
+            # we should have popped all the args, so warn if there are any left
+            logger.warning("Ignoring extra args %s from %s", func_hook.kwargs, self.description)
+
+    @property
+    def description(self):
+        return "{}:{}".format(self.plugin.title, self.function_name)
+
+    def __repr__(self):
+        return "type: {}, plugin: {}, permissions: {}, single_thread: {}, threaded: {}".format(
+            self.type, self.plugin.title, self.permissions, self.single_thread, self.threaded
+        )
