+++ cloudbot/hooks/on_stop.py
@@ -1,0 +1,12 @@
+from .hook import Hook
+
+
+class OnStopHook(Hook):
+    def __init__(self, plugin, on_stop_hook):
+        super().__init__("on_stop", plugin, on_stop_hook)
+
+    def __repr__(self):
+        return "On_stop[{}]".format(Hook.__repr__(self))
+
+    def __str__(self):
+        return "on_stop {} from {}".format(self.function_name, self.plugin.file_name)
