+++ cloudbot/hooks/periodic.py
@@ -1,0 +1,24 @@
+from .hook import Hook
+
+
+class PeriodicHook(Hook):
+    """
+    :type interval: int
+    """
+
+    def __init__(self, plugin, periodic_hook):
+        """
+        :type plugin: Plugin
+        :type periodic_hook: cloudbot.util.hook._PeriodicHook
+        """
+
+        self.interval = periodic_hook.interval
+        self.initial_interval = periodic_hook.kwargs.pop("initial_interval", self.interval)
+
+        super().__init__("periodic", plugin, periodic_hook)
+
+    def __repr__(self):
+        return "Periodic[interval: [{}], {}]".format(self.interval, Hook.__repr__(self))
+
+    def __str__(self):
+        return "periodic hook ({} seconds) {} from {}".format(self.interval, self.function_name, self.plugin.file_name)
