+++ cloudbot/hooks/post_hook.py
@@ -1,0 +1,12 @@
+from .hook import Hook
+
+
+class PostHookHook(Hook):
+    def __init__(self, plugin, out_hook):
+        super().__init__("post_hook", plugin, out_hook)
+
+    def __repr__(self):
+        return "Post_hook[{}]".format(Hook.__repr__(self))
+
+    def __str__(self):
+        return "post_hook {} from {}".format(self.function_name, self.plugin.file_name)
