+++ cloudbot/hooks/raw.py
@@ -1,0 +1,25 @@
+from .hook import Hook
+
+
+class RawHook(Hook):
+    """
+    :type triggers: set[str]
+    """
+
+    def __init__(self, plugin, irc_raw_hook):
+        """
+        :type plugin: Plugin
+        :type irc_raw_hook: cloudbot.util.hook._RawHook
+        """
+        super().__init__("irc_raw", plugin, irc_raw_hook)
+
+        self.triggers = irc_raw_hook.triggers
+
+    def is_catch_all(self):
+        return "*" in self.triggers
+
+    def __repr__(self):
+        return "Raw[triggers: {}, {}]".format(list(self.triggers), Hook.__repr__(self))
+
+    def __str__(self):
+        return "irc raw {} ({}) from {}".format(self.function_name, ",".join(self.triggers), self.plugin.file_name)
