+++ cloudbot/hooks/sieve.py
@@ -1,0 +1,16 @@
+from .hook import Hook
+
+
+class SieveHook(Hook):
+    def __init__(self, plugin, sieve_hook):
+        """
+        :type plugin: Plugin
+        :type sieve_hook: cloudbot.util.hook._SieveHook
+        """
+        super().__init__("sieve", plugin, sieve_hook)
+
+    def __repr__(self):
+        return "Sieve[{}]".format(Hook.__repr__(self))
+
+    def __str__(self):
+        return "sieve {} from {}".format(self.function_name, self.plugin.file_name)
