+++ cloudbot/plugin.py
@@ -1,6 +1,5 @@
 import asyncio
 import importlib
-import inspect
 import logging
 import sys
 from collections import defaultdict
@@ -13,7 +12,19 @@ from weakref import WeakValueDictionary
 import sqlalchemy
 
 from cloudbot.event import Event, PostHookEvent
-from cloudbot.hook import Priority, Action
+from cloudbot.hooks.cap import OnCapAvaliableHook, OnCapAckHook
+from cloudbot.hooks.command import CommandHook
+from cloudbot.hooks.event import EventHook
+from cloudbot.hooks.irc_out import IrcOutHook
+from cloudbot.hooks.on_connect import OnConnectHook
+from cloudbot.hooks.on_start import OnStartHook
+from cloudbot.hooks.on_stop import OnStopHook
+from cloudbot.hooks.periodic import PeriodicHook
+from cloudbot.hooks.permission import PermHook
+from cloudbot.hooks.post_hook import PostHookHook
+from cloudbot.hooks.raw import RawHook
+from cloudbot.hooks.regex import RegexHook
+from cloudbot.hooks.sieve import SieveHook
 from cloudbot.util import database, async_util
 from cloudbot.util.func_utils import call_with_args
 
@@ -73,12 +84,12 @@ class PluginManager:
 
     :type bot: cloudbot.bot.CloudBot
     :type plugins: dict[str, Plugin]
-    :type commands: dict[str, CommandHook]
-    :type raw_triggers: dict[str, list[RawHook]]
-    :type catch_all_triggers: list[RawHook]
-    :type event_type_hooks: dict[cloudbot.event.EventType, list[EventHook]]
-    :type regex_hooks: list[(re.__Regex, RegexHook)]
-    :type sieves: list[SieveHook]
+    :type commands: dict[str, cloudbot.hooks.command.CommandHook]
+    :type raw_triggers: dict[str, list[cloudbot.hooks.raw.RawHook]]
+    :type catch_all_triggers: list[cloudbot.hooks.raw.RawHook]
+    :type event_type_hooks: dict[cloudbot.event.EventType, list[cloudbot.hooks.event.EventHook]]
+    :type regex_hooks: list[(re.__Regex, cloudbot.hooks.regex.RegexHook)]
+    :type sieves: list[cloudbot.hooks.sieve.SieveHook]
     """
 
     def __init__(self, bot):
@@ -401,7 +412,7 @@ class PluginManager:
         """
         Logs registering a given hook
 
-        :type hook: Hook
+        :type hook: cloudbot.hooks.hook.Hook
         """
         if self.bot.config.get("logging", {}).get("show_plugin_loading", True):
             logger.info("Loaded %s", hook)
@@ -409,7 +420,7 @@ class PluginManager:
 
     def _execute_hook_threaded(self, hook, event):
         """
-        :type hook: Hook
+        :type hook: cloudbot.hooks.hook.Hook
         :type event: cloudbot.event.Event
         """
         event.prepare_threaded()
@@ -421,7 +432,7 @@ class PluginManager:
 
     async def _execute_hook_sync(self, hook, event):
         """
-        :type hook: Hook
+        :type hook: cloudbot.hooks.hook.Hook
         :type event: cloudbot.event.Event
         """
         await event.prepare()
@@ -464,7 +475,7 @@ class PluginManager:
 
         Returns False if the hook errored, True otherwise.
 
-        :type hook: cloudbot.plugin.Hook
+        :type hook: cloudbot.hooks.hook.Hook
         :type event: cloudbot.event.Event
         :rtype: bool
         """
@@ -488,9 +499,9 @@ class PluginManager:
 
     async def _sieve(self, sieve, event, hook):
         """
-        :type sieve: cloudbot.plugin.Hook
+        :type sieve: cloudbot.hooks.hook.Hook
         :type event: cloudbot.event.Event
-        :type hook: cloudbot.plugin.Hook
+        :type hook: cloudbot.hooks.hook.Hook
         :rtype: cloudbot.event.Event
         """
         if sieve.threaded:
@@ -537,7 +548,7 @@ class PluginManager:
         Returns False if the hook didn't run successfully, and True if it ran successfully.
 
         :type event: cloudbot.event.Event | cloudbot.event.CommandEvent
-        :type hook: cloudbot.plugin.Hook | cloudbot.plugin.CommandHook
+        :type hook: cloudbot.hooks.hook.Hook | cloudbot.hooks.command.CommandHook
         :rtype: bool
         """
 
@@ -642,303 +653,6 @@ class Plugin:
 
             for table in self.tables:
                 bot.db_metadata.remove(table)
-
-
-class Hook:
-    """
-    Each hook is specific to one function. This class is never used by itself, rather extended.
-
-    :type type; str
-    :type plugin: Plugin
-    :type function: callable
-    :type function_name: str
-    :type required_args: list[str]
-    :type threaded: bool
-    :type permissions: list[str]
-    :type single_thread: bool
-    """
-
-    def __init__(self, _type, plugin, func_hook):
-        """
-        :type _type: str
-        :type plugin: Plugin
-        :type func_hook: hook._Hook
-        """
-        self.type = _type
-        self.plugin = plugin
-        self.function = func_hook.function
-        self.function_name = self.function.__name__
-
-        sig = inspect.signature(self.function)
-
-        # don't process args starting with "_"
-        self.required_args = [arg for arg in sig.parameters.keys() if not arg.startswith('_')]
-
-        if asyncio.iscoroutine(self.function) or asyncio.iscoroutinefunction(self.function):
-            self.threaded = False
-        else:
-            self.threaded = True
-
-        self.permissions = func_hook.kwargs.pop("permissions", [])
-        self.single_thread = func_hook.kwargs.pop("singlethread", False)
-        self.action = func_hook.kwargs.pop("action", Action.CONTINUE)
-        self.priority = func_hook.kwargs.pop("priority", Priority.NORMAL)
-
-        clients = func_hook.kwargs.pop("clients", [])
-
-        if isinstance(clients, str):
-            clients = [clients]
-
-        self.clients = clients
-
-        if func_hook.kwargs:
-            # we should have popped all the args, so warn if there are any left
-            logger.warning("Ignoring extra args %s from %s", func_hook.kwargs, self.description)
-
-    @property
-    def description(self):
-        return "{}:{}".format(self.plugin.title, self.function_name)
-
-    def __repr__(self):
-        return "type: {}, plugin: {}, permissions: {}, single_thread: {}, threaded: {}".format(
-            self.type, self.plugin.title, self.permissions, self.single_thread, self.threaded
-        )
-
-
-class CommandHook(Hook):
-    """
-    :type name: str
-    :type aliases: list[str]
-    :type doc: str
-    :type auto_help: bool
-    """
-
-    def __init__(self, plugin, cmd_hook):
-        """
-        :type plugin: Plugin
-        :type cmd_hook: cloudbot.util.hook._CommandHook
-        """
-        self.auto_help = cmd_hook.kwargs.pop("autohelp", True)
-
-        self.name = cmd_hook.main_alias.lower()
-        self.aliases = [alias.lower() for alias in cmd_hook.aliases]  # turn the set into a list
-        self.aliases.remove(self.name)
-        self.aliases.insert(0, self.name)  # make sure the name, or 'main alias' is in position 0
-        self.doc = cmd_hook.doc
-
-        super().__init__("command", plugin, cmd_hook)
-
-    def __repr__(self):
-        return "Command[name: {}, aliases: {}, {}]".format(self.name, self.aliases[1:], Hook.__repr__(self))
-
-    def __str__(self):
-        return "command {} from {}".format("/".join(self.aliases), self.plugin.file_name)
-
-
-class RegexHook(Hook):
-    """
-    :type regexes: set[re.__Regex]
-    """
-
-    def __init__(self, plugin, regex_hook):
-        """
-        :type plugin: Plugin
-        :type regex_hook: cloudbot.util.hook._RegexHook
-        """
-        self.run_on_cmd = regex_hook.kwargs.pop("run_on_cmd", False)
-        self.only_no_match = regex_hook.kwargs.pop("only_no_match", False)
-
-        self.regexes = regex_hook.regexes
-
-        super().__init__("regex", plugin, regex_hook)
-
-    def __repr__(self):
-        return "Regex[regexes: [{}], {}]".format(", ".join(regex.pattern for regex in self.regexes),
-                                                 Hook.__repr__(self))
-
-    def __str__(self):
-        return "regex {} from {}".format(self.function_name, self.plugin.file_name)
-
-
-class PeriodicHook(Hook):
-    """
-    :type interval: int
-    """
-
-    def __init__(self, plugin, periodic_hook):
-        """
-        :type plugin: Plugin
-        :type periodic_hook: cloudbot.util.hook._PeriodicHook
-        """
-
-        self.interval = periodic_hook.interval
-        self.initial_interval = periodic_hook.kwargs.pop("initial_interval", self.interval)
-
-        super().__init__("periodic", plugin, periodic_hook)
-
-    def __repr__(self):
-        return "Periodic[interval: [{}], {}]".format(self.interval, Hook.__repr__(self))
-
-    def __str__(self):
-        return "periodic hook ({} seconds) {} from {}".format(self.interval, self.function_name, self.plugin.file_name)
-
-
-class RawHook(Hook):
-    """
-    :type triggers: set[str]
-    """
-
-    def __init__(self, plugin, irc_raw_hook):
-        """
-        :type plugin: Plugin
-        :type irc_raw_hook: cloudbot.util.hook._RawHook
-        """
-        super().__init__("irc_raw", plugin, irc_raw_hook)
-
-        self.triggers = irc_raw_hook.triggers
-
-    def is_catch_all(self):
-        return "*" in self.triggers
-
-    def __repr__(self):
-        return "Raw[triggers: {}, {}]".format(list(self.triggers), Hook.__repr__(self))
-
-    def __str__(self):
-        return "irc raw {} ({}) from {}".format(self.function_name, ",".join(self.triggers), self.plugin.file_name)
-
-
-class SieveHook(Hook):
-    def __init__(self, plugin, sieve_hook):
-        """
-        :type plugin: Plugin
-        :type sieve_hook: cloudbot.util.hook._SieveHook
-        """
-        super().__init__("sieve", plugin, sieve_hook)
-
-    def __repr__(self):
-        return "Sieve[{}]".format(Hook.__repr__(self))
-
-    def __str__(self):
-        return "sieve {} from {}".format(self.function_name, self.plugin.file_name)
-
-
-class EventHook(Hook):
-    """
-    :type types: set[cloudbot.event.EventType]
-    """
-
-    def __init__(self, plugin, event_hook):
-        """
-        :type plugin: Plugin
-        :type event_hook: cloudbot.util.hook._EventHook
-        """
-        super().__init__("event", plugin, event_hook)
-
-        self.types = event_hook.types
-
-    def __repr__(self):
-        return "Event[types: {}, {}]".format(list(self.types), Hook.__repr__(self))
-
-    def __str__(self):
-        return "event {} ({}) from {}".format(self.function_name, ",".join(str(t) for t in self.types),
-                                              self.plugin.file_name)
-
-
-class OnStartHook(Hook):
-    def __init__(self, plugin, on_start_hook):
-        """
-        :type plugin: Plugin
-        :type on_start_hook: cloudbot.util.hook._On_startHook
-        """
-        super().__init__("on_start", plugin, on_start_hook)
-
-    def __repr__(self):
-        return "On_start[{}]".format(Hook.__repr__(self))
-
-    def __str__(self):
-        return "on_start {} from {}".format(self.function_name, self.plugin.file_name)
-
-
-class OnStopHook(Hook):
-    def __init__(self, plugin, on_stop_hook):
-        super().__init__("on_stop", plugin, on_stop_hook)
-
-    def __repr__(self):
-        return "On_stop[{}]".format(Hook.__repr__(self))
-
-    def __str__(self):
-        return "on_stop {} from {}".format(self.function_name, self.plugin.file_name)
-
-
-class CapHook(Hook):
-    def __init__(self, _type, plugin, base_hook):
-        self.caps = base_hook.caps
-        super().__init__("on_cap_{}".format(_type), plugin, base_hook)
-
-    def __repr__(self):
-        return "{name}[{caps} {base!r}]".format(name=self.type, caps=self.caps, base=super())
-
-    def __str__(self):
-        return "{name} {func} from {file}".format(name=self.type, func=self.function_name, file=self.plugin.file_name)
-
-
-class OnCapAvaliableHook(CapHook):
-    def __init__(self, plugin, base_hook):
-        super().__init__("available", plugin, base_hook)
-
-
-class OnCapAckHook(CapHook):
-    def __init__(self, plugin, base_hook):
-        super().__init__("ack", plugin, base_hook)
-
-
-class OnConnectHook(Hook):
-    def __init__(self, plugin, sieve_hook):
-        """
-        :type plugin: Plugin
-        :type sieve_hook: cloudbot.util.hook._Hook
-        """
-        super().__init__("on_connect", plugin, sieve_hook)
-
-    def __repr__(self):
-        return "{name}[{base!r}]".format(name=self.type, base=super())
-
-    def __str__(self):
-        return "{name} {func} from {file}".format(name=self.type, func=self.function_name, file=self.plugin.file_name)
-
-
-class IrcOutHook(Hook):
-    def __init__(self, plugin, out_hook):
-        super().__init__("irc_out", plugin, out_hook)
-
-    def __repr__(self):
-        return "Irc_Out[{}]".format(Hook.__repr__(self))
-
-    def __str__(self):
-        return "irc_out {} from {}".format(self.function_name, self.plugin.file_name)
-
-
-class PostHookHook(Hook):
-    def __init__(self, plugin, out_hook):
-        super().__init__("post_hook", plugin, out_hook)
-
-    def __repr__(self):
-        return "Post_hook[{}]".format(Hook.__repr__(self))
-
-    def __str__(self):
-        return "post_hook {} from {}".format(self.function_name, self.plugin.file_name)
-
-
-class PermHook(Hook):
-    def __init__(self, plugin, perm_hook):
-        self.perms = perm_hook.perms
-        super().__init__("perm_check", plugin, perm_hook)
-
-    def __repr__(self):
-        return "PermHook[{}]".format(Hook.__repr__(self))
-
-    def __str__(self):
-        return "perm hook {} from {}".format(self.function_name, self.plugin.file_name)
 
 
 _hook_name_to_plugin = {
