+++ tests/core_tests/test_plugin_hooks.py
@@ -12,9 +12,10 @@ from pathlib import Path
 import pytest
 
 import cloudbot.bot
-from cloudbot.event import Event, CommandEvent, RegexEvent, CapEvent, PostHookEvent, IrcOutEvent
+from cloudbot.event import CapEvent, CommandEvent, Event, IrcOutEvent, PostHookEvent, RegexEvent
 from cloudbot.hook import Action
-from cloudbot.plugin import Plugin, Hook
+from cloudbot.hooks.hook import Hook
+from cloudbot.plugin import Plugin
 
 Hook.original_init = Hook.__init__
